requirements.txt
docker==7.0.0
kubernetes==29.0.0
rich==13.7.0
python-dotenv==1.0.1
click==8.1.7
requests==2.31.0
pyyaml==6.0.1 